review: move QueryEnum + mark search as deprecated
