+++ lib/collection/src/shards/local_shard/search.rs
@@ -3,12 +3,12 @@ use std::time::Duration;
 
 use common::counter::hardware_accumulator::HwMeasurementAcc;
 use segment::types::ScoredPoint;
+use shard::query::query_enum::QueryEnum;
 use tokio::runtime::Handle;
 
 use super::LocalShard;
 use crate::collection_manager::segments_searcher::SegmentsSearcher;
 use crate::common::stopping_guard::StoppingGuard;
-use crate::operations::query_enum::QueryEnum;
 use crate::operations::types::{CollectionError, CollectionResult, CoreSearchRequestBatch};
 
 // Chunk requests for parallelism in certain scenarios
