+++ lib/shard/src/query/mod.rs
@@ -1,0 +1,1 @@
+pub mod query_enum;
